refactor: refactored code and files
