+++ src/api-calls/index.js
@@ -1,0 +1,4 @@
+import { getProductFromServer, getProductByIdfromServer, getTypeCategoriesFromServer } from "./getMethodApiCalls";
+import { signInHandler, loginHandler } from "./postMethodApiCalls";
+
+export {getProductByIdfromServer, getProductFromServer, getTypeCategoriesFromServer, signInHandler, loginHandler};
